Include contrib in webpack-dev-server
+ Moved common configurations to webpack.common.js
