+++ contrib/auto-render/index.html
@@ -1,15 +1,14 @@
 <!DOCTYPE html>
-<!--To run this example from a clone of the repository, run `make serve`
+<!--To run this example from a clone of the repository, run `npm start`
     in the root KaTeX directory and then visit with your web browser:
-    http://0.0.0.0:7936/contrib/auto-render/index.html
+    http://localhost:7936/contrib/auto-render/index.html
   -->
 <html>
   <head>
     <meta charset="UTF-8">
     <title>Auto-render test</title>
     <script src="/katex.js" type="text/javascript"></script>
-    <link href="/katex.css" rel="stylesheet" type="text/css">
-    <script src="./auto-render.js" type="text/javascript"></script>
+    <script src="/contrib/auto-render.js" type="text/javascript"></script>
     <style type="text/css">
       body {
           margin: 0px;
