+++ contrib/copy-tex/index.html
@@ -1,17 +1,15 @@
 <!DOCTYPE html>
-<!--To run this example from a clone of the repository, run `make serve`
+<!--To run this example from a clone of the repository, run `npm start`
     in the root KaTeX directory and then visit with your web browser:
-    http://0.0.0.0:7936/contrib/copy-tex/index.html
+    http://localhost:7936/contrib/copy-tex/index.html
   -->
 <html>
   <head>
     <meta charset="UTF-8">
     <title>Copy-tex test</title>
     <script src="/katex.js" type="text/javascript"></script>
-    <link href="/katex.css" rel="stylesheet" type="text/css">
-    <link href="./copy-tex.css" rel="stylesheet" type="text/css">
-    <script src="../auto-render/auto-render.js" type="text/javascript"></script>
-    <script src="./copy-tex.js" type="text/javascript"></script>
+    <script src="/contrib/auto-render.js" type="text/javascript"></script>
+    <script src="/contrib/copy-tex.js" type="text/javascript"></script>
     <style type="text/css">
       body {
           margin: 0px;
