+++ dockers/Screenshotter/screenshotter.js
@@ -196,13 +196,8 @@ function startServer() {
     }
     const port = Math.floor(Math.random() * (maxPort - minPort)) + minPort;
     const compiler = webpack(webpackConfig);
-    const server = new webpackDevServer(compiler, {
-        contentBase: [
-            path.join(__dirname, '../../static'),
-            path.join(__dirname, '../..'),
-        ],
-        disableHostCheck: true,
-    }).listen(port);
+    const server = new webpackDevServer(compiler,
+        webpackConfig[0].devServer).listen(port);
     server.once("listening", function() {
         devServer = server;
         katexPort = port;
